build api wip

+++ src/routes/api/host.js
@@ -13,6 +13,30 @@ const Host = require('../../models/host');
 /**
  * @swagger
  * /host:
+ *  get:
+ *    tags:
+ *      - host
+ *    description: List all docker hosts
+ *    produces:
+ *      - application/json
+ *    responses:
+ *      200:
+ *        description: All docker hosts serialized objects
+ *      500:
+ *        description: Internal server error
+ *
+ */
+
+router.get('/host', async (req, res, next) => {
+    const hosts = await Host.find();
+
+    res.status(200);
+    res.send(hosts.map(host => host.serialized()));
+});
+
+/**
+ * @swagger
+ * /host:
  *  post:
  *    tags:
  *      - host
@@ -23,10 +47,14 @@ const Host = require('../../models/host');
  *    - name: url
  *      in: formData
  *      required: false
- *      schema:
- *         type: string
- *         default: '/var/run/docker.sock'
+ *      type: string
+ *      default: '/var/run/docker.sock'
  *      description: The docker host url, defaults to '/var/run/docker.sock'
+ *    - name: ca
+ *      in: formData
+ *      required: false
+ *      type: file
+ *      description: The docker host ssl certificate authority path
  *    - name: cert
  *      in: formData
  *      required: false
@@ -49,21 +77,35 @@ const Host = require('../../models/host');
  *
  */
 
-router.post('/host', upload.fields([{ name: 'cert' }, { name: 'key' }]), async (req, res, next) => {
+router.post('/host', upload.fields([{ name: 'ca' }, { name: 'cert' }, { name: 'key' }]), async (req, res, next) => {
     let errors = [];
     // set url property, defaults to unix docker socket /var/run/docker.sock
     const url = void 0 === req.body.url ? config.dockerSocketPath : req.body.url;
     let host = {
         url: url
-    }
+    };
 
     // cancel if docker socket has already been registered
     if (0 < (await Host.find({url: url})).length) {
         errors.push('this docker socket/url has already been registered');
     }
 
+    try {
+        fs.accessSync(config.uploadPath, fs.constants.F_OK);
+    } catch (err) {
+        console.error(err);
+        fs.mkdirSync(config.uploadPath, {'mode': 0o775});
+    }
+
     // copy then delete uploaded files from /tmp to config.uploadPath directory
-    if (req.files.cert || req.files.key) {
+    if (req.files.ca || req.files.cert || req.files.key) {
+        fs.copyFile(req.files.ca[0].path, config.uploadPath+req.files.ca[0].filename, (err) => {
+            if (err) serverError(res, err);
+        });
+        fs.unlink(req.files.ca[0].path, (err) => {
+            if (err) serverError(res, err);
+        });
+
         fs.copyFile(req.files.cert[0].path, config.uploadPath+req.files.cert[0].filename, (err) => {
             if (err) serverError(res, err);
         });
@@ -81,13 +123,16 @@ router.post('/host', upload.fields([{ name: 'cert' }, { name: 'key' }]), async (
 
     // cancel if there is certificate and key files but no valid url
     const regex = /(?<host>^.+):(?<port>\d+$)/;
-    if (req.files.cert && req.files.key && !regex.test(url)) {
-        errors.push('url does not match host:port pattern while providing certs');
+    if (req.files.ca && req.files.cert && req.files.key && !regex.test(url)) {
+        errors.push('url does not match host:port pattern while providing certificates');
     }
 
     // remove files if there is at least an error
     if (0 < errors.length) {
-        if (req.files.cert && req.files.key) {
+        if (req.files.ca && req.files.cert && req.files.key) {
+            fs.unlink(config.uploadPath+req.files.ca[0].filename, (err) => {
+                if (err) serverError(res, err);
+            });
             fs.unlink(config.uploadPath+req.files.cert[0].filename, (err) => {
                 if (err) serverError(res, err);
             });
@@ -105,7 +150,8 @@ router.post('/host', upload.fields([{ name: 'cert' }, { name: 'key' }]), async (
     }
 
     // set cert & key properties
-    if (req.files.cert && req.files.key && regex.test(url)) {
+    if (req.files.ca && req.files.cert && req.files.key && regex.test(url)) {
+        host.ca = req.files.ca[0].filename;
         host.cert = req.files.cert[0].filename;
         host.key = req.files.key[0].filename;
     }
@@ -116,31 +162,7 @@ router.post('/host', upload.fields([{ name: 'cert' }, { name: 'key' }]), async (
     await host.save();
 
      res.status(200);
-     res.send(host);
-});
-
-/**
- * @swagger
- * /host:
- *  get:
- *    tags:
- *      - host
- *    description: Read all docker host
- *    produces:
- *      - application/json
- *    responses:
- *      200:
- *        description: All docker hosts serialized objects
- *      500:
- *        description: Internal server error
- *
- */
-
-router.get('/host', async (req, res, next) => {
-    const hosts = await Host.find();
-
-    res.status(200);
-    res.send(hosts);
+     res.send(host.serialized());
 });
 
 /**
@@ -149,7 +171,7 @@ router.get('/host', async (req, res, next) => {
  *  get:
  *    tags:
  *      - host
- *    description: Read a docker host
+ *    description: Return a docker host
  *    parameters:
  *    - name: id
  *      in: path
@@ -169,7 +191,7 @@ router.get('/host', async (req, res, next) => {
  */
 
 router.get('/host/:id', async (req, res, next) => {
-    const host = await Host.findOne({id: req.param.id});
+    const host = await Host.findOne({_id: req.params.id});
 
     if (!host) {
         res.status(404);
@@ -180,7 +202,7 @@ router.get('/host/:id', async (req, res, next) => {
     }
 
     res.status(200);
-    res.send(host);
+    res.send(host.serialized());
 });
 
 /**
@@ -209,7 +231,7 @@ router.get('/host/:id', async (req, res, next) => {
  */
 
 router.delete('/host/:id', async (req, res, next) => {
-    const host = await Host.findOne({id: req.param.id});
+    const host = await Host.findOne({_id: req.params.id});
 
     if (!host) {
         res.status(404);
@@ -219,7 +241,10 @@ router.delete('/host/:id', async (req, res, next) => {
         throw httpError('404', 'Resource not found');
     }
 
-    if (host.cert && host.key) {
+    if (host.ca && host.cert && host.key) {
+        fs.unlink(config.uploadPath+host.ca, (err) => {
+            if (err) serverError(res, err);
+        });
         fs.unlink(config.uploadPath+host.cert, (err) => {
             if (err) serverError(res, err);
         });
@@ -231,7 +256,7 @@ router.delete('/host/:id', async (req, res, next) => {
     await Host.deleteOne({_id: host.id });
 
     res.status(200);
-    res.send(host);
+    res.send(host.serialized());
 });
 
 module.exports = router;
